change file name from server.rs to room_manager.rs and refactor
