+++ src/websocket/room_manager.rs
